// src/signs/generated/artwork/D1-1.ts
// Generated from official FHWA 2024 SHS artwork. Do not edit.
import type { SignArt } from "../../artwork";

export const ART_D1_1: SignArt = {
  "width": 9,
  "height": 6,
  "paths": [
    {
      "d": "M9.002 4.098L3.986 4.098L3.986 1.848L9.002 1.848Z",
      "fill": "white"
    },
    {
      "d": "M4.862 5.492C4.872 5.527 4.877 5.563 4.877 5.597C4.877 5.804 4.709 5.972 4.502 5.972C4.428 5.972 4.355 5.951 4.29 5.907L0.002 2.972L4.29 0.038C4.355 -0.007 4.428 -0.028 4.502 -0.028C4.709 -0.028 4.877 0.14 4.877 0.347C4.877 0.382 4.872 0.417 4.862 0.452L4.127 2.972Z",
      "fill": "white"
    }
  ]
};