// src/signs/generated/artwork/E5-1.ts
// Generated from official FHWA 2024 SHS artwork. Do not edit.
import type { SignArt } from "../../artwork";

export const ART_E5_1: SignArt = {
  "width": 23,
  "height": 23,
  "paths": [
    {
      "d": "M22.78 0.22C22.598 0.037 22.345 -0.031 22.11 0.013L7.354 2.929C7.209 2.956 7.071 3.025 6.959 3.136C6.666 3.429 6.666 3.904 6.959 4.197C7.007 4.245 7.06 4.285 7.116 4.318L12.351 7.467L-0.024 18.781L4.219 23.024L15.533 10.649L18.682 15.884C18.715 15.94 18.755 15.993 18.803 16.041C19.096 16.334 19.571 16.334 19.864 16.041C19.975 15.929 20.044 15.791 20.071 15.646L22.987 0.89C23.031 0.655 22.963 0.402 22.78 0.22Z",
      "fill": "white"
    }
  ]
};